backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-socketio>=5.10.0
python-multipart>=0.0.6
httpx>=0.25.2
redis>=5.0.1
pydantic>=2.7.4
python-dotenv>=1.0.0
google-generativeai>=0.3.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
